Move advanced tab params into separate file
