+++ rtengine/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(PROCPARAM_SRCS
     procparams.cc
+    params/advanced.cc
     params/locallab.cc
     params/raw.cc
     params/serdes.cc
